+++ energizer/learners/base.py
@@ -5,13 +5,14 @@ from pytorch_lightning import LightningModule
 from torch import Tensor
 
 from energizer.learners.hooks import ModelHooks
-from energizer.utilities.learner_utils import patch_dropout_layers
+from energizer.utilities.mcdropout_utils import patch_dropout_layers
 
 
 class PostInitCaller(type):
     """Used to call `setup` automatically after `__init__`"""
+
     def __call__(cls, *args, **kwargs):
-        """Called when you call MyNewClass() """
+        """Called when you call MyNewClass()"""
         obj = type.__call__(cls, *args, **kwargs)
         obj.__post_init__()
         return obj
@@ -32,10 +33,10 @@ class Learner(LightningModule, ModelHooks, metaclass=PostInitCaller):
 
     def pool_epoch_end(self, *args, **kwargs) -> Optional[Any]:
         pass
-    
+
     def _forward(self, *args, **kwargs) -> Any:
         return self.forward(*args, **kwargs)
-    
+
     def __call__(self, *args, **kwargs) -> Any:
         return self._forward(*args, **kwargs)
 
@@ -51,7 +52,6 @@ class MCDropoutMixin(Learner):
         self,
         num_inference_iters: Optional[int] = 10,
         consistent: Optional[bool] = False,
-        inplace: Optional[bool] = True,
         prob: Optional[float] = None,
     ) -> None:
         """Instantiates a new learner (same as `learner`) with patched dropout layers.
@@ -68,7 +68,6 @@ class MCDropoutMixin(Learner):
         self.num_inference_iters = num_inference_iters
         self.consistent = consistent
         self.prob = prob
-        self.inplace = inplace
         super().__init__()
 
     def __post_init__(self) -> None:
@@ -76,7 +75,7 @@ class MCDropoutMixin(Learner):
             module=self,
             prob=self.prob,
             consistent=self.consistent,
-            inplace=self.inplace,
+            inplace=True,
         )
 
     def _forward(self, *args, **kwargs) -> Tensor:
